+++ .ncurc.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  reject: [
+    // Wait until making this package ESM and ESM-only
+    'node-fetch'
+  ]
+};